Add missing xolite.html
I also changed the way we select files from the archive: now we copy
everything so that nothing is missed, but still mention the files
individually in the %files section. Thus, if files are added in the
future, we'll notice it immediately because rpmbuild will complain about
the new unknown files.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>